+++ src/components/functional/NavigateToSurvey.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { getCountAndTimeLeft } from "../services/countService";
-import { useAuth } from "../contexts/AuthContext";
+import { getCountAndTimeLeft } from "../../services/countService";
+import { useAuth } from "../../contexts/AuthContext";
 import { Button } from "./Button";
 
 interface NavigateToSurveyProps {
