+++ src/pages/Home.tsx
@@ -3,12 +3,12 @@ import { useNavigate } from "react-router-dom";
 import { signInWithEmailAndPassword } from "firebase/auth";
 import { auth } from "../firebase";
 import { Button } from "../components/Button"
-import { Main } from "../components/Main";
-import Input from "../components/Input";
+import { Main } from "../components/ui/Main";
+import Input from "../components/ui/Input";
 import { useAuth } from "../contexts/AuthContext";
 import { ButtonGroup } from "../styles/styled";
-import { FlexBox } from "../components/FlexBox";
-import { Text } from "../components/Text";
+import { FlexBox } from "../components/ui/FlexBox";
+import { Text } from "../components/ui/Text";
 import { FirebaseError } from "firebase/app";
 import { useResponsive } from "../hooks/useResponsive";
